Moved python category code to main crate. Added some comparison ops.
