+++ rpycocotools/src/lib.rs
@@ -7,7 +7,7 @@ pub mod errors;
 #[pymodule]
 fn rpycocotools(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_class::<coco::PyCOCO>()?;
-    m.add_class::<coco::PyCategory>()?;
+    m.add_class::<cocotools::annotations::coco::Annotation>()?;
     // m.add_function(wrap_pyfunction!(visualize_img, m)?)?;
     Ok(())
 }
